feat: remove version.py; expose __version__ from package metadata with pyproject fallback; update imports to package-level __version__; update tests; ruff fix/format and tests pass
